+++ MyBudgetExplorer/Pages/Budget/Account.cshtml.cs
@@ -84,7 +84,7 @@ namespace MyBudgetExplorer.Pages.Budget
                 Transactions.Add(trans);
             }
 
-            var currentBalance = 0;
+            var currentBalance = 0L;
             foreach (var t in Transactions)
             {
                 currentBalance += t.Amount;
@@ -98,7 +98,7 @@ namespace MyBudgetExplorer.Pages.Budget
             public DateTime Date { get; set; }
             public string DateDisplay { get { return Date.ToShortDateString(); } }
             public string Payee { get; set; }
-            public int Amount { get; set; }
+            public long Amount { get; set; }
             public string AmountDisplay { get { return Amount.ToDisplay(); } }
             public string Available { get; set; }
         }
